airflow/requirements.txt
requests
apache-airflow
python-dotenv
pandas
pyarrow
psycopg2-binary
boto3

